refactor(settings): decompose SettingsPage into focused section components
Split the 498-line SettingsPage into 8 focused section components:
- ProfileSection: user profile display with occupations
- LanguageSection: language switcher
- TourSection: guided tour status and reset
- DemoSection: demo data reset functionality
- SafeModeSection: safe mode toggle with warning modal
- PrivacySection: privacy information display
- UpdateSection: PWA update checks
- AboutSection: app version and links

This improves maintainability by separating concerns and reducing
the main page component to ~88 lines.
