+++ web-app/src/components/features/settings/index.ts
@@ -1,0 +1,8 @@
+export { ProfileSection } from "./ProfileSection";
+export { LanguageSection } from "./LanguageSection";
+export { TourSection } from "./TourSection";
+export { DemoSection } from "./DemoSection";
+export { SafeModeSection } from "./SafeModeSection";
+export { PrivacySection } from "./PrivacySection";
+export { UpdateSection } from "./UpdateSection";
+export { AboutSection } from "./AboutSection";
